feat: implement level object parsing, model metadata loading, and QSC serialization logic

+++ source/level/level_objects.h
@@ -9,8 +9,8 @@ struct LevelObject {
     std::string name;
     std::string modelId;
     std::string taskId;  // Task ID from QSC for save support
-    glm::vec3 pos;
-    glm::vec3 rot;
+    glm::dvec3 pos;
+    glm::dvec3 rot;
     bool isBuilding;
 
     // Lighting
